Use proper non-deterministic temporary directories
This makes sure that the temporary directory name cannot be predicted,
to do that we also need to create one per package, as we are removing
them before proceeding to the next one.

By using the File::Temp object interface we also get automatic cleanup
on exit, w/o the need to track it ourselves.

This fixes a possible security issue, and using --generate on multiple
packages.