Use provided pkgname instead of partially reconstructing it from metadata
When we get a package name, we should use exactly that for all
dpkg-query calls that expect a package name, because otherwise we
might get into situations were we should or should not arch-qualify
it but have not done so.

We could always arch-qualify the package names but that would mean
we cannot support older dpkg versions.

Closes: #870724