debian: Replace experimental usage of «/» in build profile
This was done to check whether it was currently feasible to use such
characters, even though currently not permitted by the spec, but do
fix a design flaw, given that package names can contain «.» which is
the current designated namespace separator.

It looks like some infrastructure will need to be fixed before this can
be reconsidered.

Closes: #1003814